LongTask Timing_ long task in cross-origin sibling iframe.html

<!DOCTYPE html>
<!-- saved from url=(0079)http://w3c-test.org/longtask-timing/longtask-in-sibling-iframe-crossorigin.html -->
<html><head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
<title>LongTask Timing: long task in cross-origin sibling iframe</title>
<style>html {    font-family:DejaVu Sans, Bitstream Vera Sans, Arial, Sans;}#log .warning,#log .warning a {  color: black;  background: yellow;}#log .error,#log .error a {  color: white;  background: red;}section#summary {    margin-bottom:1em;}table#results {    border-collapse:collapse;    table-layout:fixed;    width:100%;}table#results > thead > tr > th:first-child,table#results > tbody > tr > td:first-child {    width:8em;}table#results > thead > tr > th:last-child,table#results > thead > tr > td:last-child {    width:50%;}table#results.assertions > thead > tr > th:last-child,table#results.assertions > tbody > tr > td:last-child {    width:35%;}table#results > thead > > tr > th {    padding:0;    padding-bottom:0.5em;    border-bottom:medium solid black;}table#results > tbody > tr> td {    padding:1em;    padding-bottom:0.5em;    border-bottom:thin solid black;}.pass {    color:green;}.fail {    color:red;}tr.timeout {    color:red;}tr.notrun {    color:blue;}tr.optionalunsupported {    color:blue;}.ok {    color:green;}.error {    color:red;}.pass, .fail, .timeout, .notrun, .optionalunsupported .ok, .timeout, .error {    font-variant:small-caps;}table#results span {    display:block;}table#results span.expected {    font-family:DejaVu Sans Mono, Bitstream Vera Sans Mono, Monospace;    white-space:pre;}table#results span.actual {    font-family:DejaVu Sans Mono, Bitstream Vera Sans Mono, Monospace;    white-space:pre;}</style></head><body>

<script src="./LongTask Timing_ long task in cross-origin sibling iframe_files/testharness.js.download"></script>
<script src="./LongTask Timing_ long task in cross-origin sibling iframe_files/testharnessreport.js.download"></script>

<script>
  async_test(t => {
    assert_implements(window.PerformanceLongTaskTiming, 'Longtasks are not supported.');
    window.addEventListener('message', t.step_func(e => {
      assert_equals(e.data['entryType'], 'longtask');
      assert_equals(e.data['frame-attribution'], 'cross-origin-unreachable');
      assert_equals(e.data['task-attribution'], 'unknown');
      assert_equals(e.data['containerType'], 'window');
      assert_equals(e.data['containerId'], '');
      assert_equals(e.data['containerName'], '');
      assert_equals(e.data['containerSrc'], '');
      t.done();
    }));
    const observingFrame = document.createElement('iframe');
    observingFrame.id = 'observing-iframe-id';
    observingFrame.name = 'observing-iframe-name';
    document.body.appendChild(observingFrame);
    observingFrame.src = 'resources/subframe-observing-longtask.html'

    observingFrame.onload = () => {
      /* Create a cross-origin iframe that generates a long task. */
      const longtaskFrame = document.createElement('iframe');
      longtaskFrame.id = 'longtask-iframe-id';
      longtaskFrame.name = 'longtask-iframe-name';
      // Simulate cross-origin by using sandbox.
      longtaskFrame.sandbox = "allow-scripts";
      document.body.appendChild(longtaskFrame);
      longtaskFrame.src = 'resources/subframe-with-longtask.html'
    };
}, 'Performance longtask entries from cross-origin iframe are observable in its sibling.');
</script><iframe id="observing-iframe-id" name="observing-iframe-name" src="./LongTask Timing_ long task in cross-origin sibling iframe_files/subframe-observing-longtask.html"></iframe>

<iframe id="longtask-iframe-id" name="longtask-iframe-name" sandbox="allow-scripts" src="./LongTask Timing_ long task in cross-origin sibling iframe_files/subframe-with-longtask.html"></iframe><div id="log"><section id="summary"><h2>Summary</h2><section><p>Harness status: <span class="ok">OK</span></p></section><p>Found 1 tests</p><div><div><label><input type="checkbox" checked="checked">1 <span class="pass">Pass</span></label></div></div></section><section><h2>Details</h2><table id="results"><thead><tr><th>Result</th><th>Test Name</th><th>Message</th></tr></thead><tbody><tr class="overall-pass"><td class="pass">Pass</td><td>Performance longtask entries from cross-origin iframe are observable in its sibling.</td><td> <details><summary>Asserts run</summary><table><tbody><tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_implements</strong>(function "function PerformanceLongTaskTiming() { [native code] }", "Longtasks are not supported.")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-sibling-iframe-crossorigin.html:11:5)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("longtask", "longtask")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-sibling-iframe-crossorigin.html:13:7)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("cross-origin-unreachable", "cross-origin-unreachable")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-sibling-iframe-crossorigin.html:14:7)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("unknown", "unknown")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-sibling-iframe-crossorigin.html:15:7)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("window", "window")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-sibling-iframe-crossorigin.html:16:7)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("", "")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-sibling-iframe-crossorigin.html:17:7)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("", "")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-sibling-iframe-crossorigin.html:18:7)</pre></td></tr>
<tr class="overall-pass"><td class="pass">Pass</td><td><pre><strong>assert_equals</strong>("", "")
    at Test.&lt;anonymous&gt; ( /longtask-timing/longtask-in-sibling-iframe-crossorigin.html:19:7)</pre></td></tr></tbody></table></details></td></tr></tbody></table></section></div><script type="text/json" id="__testharness__results__">{"test":"http://w3c-test.org/longtask-timing/longtask-in-sibling-iframe-crossorigin.html","tests":[{"name":"Performance longtask entries from cross-origin iframe are observable in its sibling.","status":0,"message":null,"stack":null}],"status":0,"message":null,"stack":null}</script></body></html>
